Successfully separated resources from the code
All the resources (routes) are now placed inside the resources directory.
The application class is kept; created a separated method, create_app(), that
only creates the application and passes all the resources to the Flask api.

Restructured some of the files.
The comments are not up-to-date. Will update them accordingly after
making sure the architecture works fine and nothing is change regarding the
fundamental logic.