Applied schema validation
Created schema fields and applied validation.
The POST requests seems to work fine and it's pretty quick (< 0.1 sec. from my device).
Now comes the hard part. I am finally able to get the neural network model and
call it inside the request. Will need to load it in some kind of cache
such that I don't have to construct the weights every time a prediction
need to be made.