Updated FrameReceiver
The model actually works. Used ffmpeg and h264 compression algorithm to reduce the
size of the .mp4 file and decided to send it through POST method.
The FrameReceiver loads the file and converts it to numpy array. After that,
it applied neural network model prediction.

Also, cached the NNAdapter (which initializes the model and loads the weights)
using __init__() method. This significantly reduces execution time (from ~7 seconds to ~1 second).